QFBNLEnv in working order
This environment takes into consideration the current rate limiting on each quadrupole, implements the PI controller in use in the QFB with default settings and can also perturb the state at S_{t+1} (in the standard RL nomenclature) to behave like the tunes as estimated by the BQ algorithm on real BBQ spectra. This was done by simulating a dummy spectrum with injected 50 Hz harmonics, and applying the BQ algorithm to said spectrum. The resulting tune estimate is then provided to the agent.